:recycle: Home component
added Home component on index
